requirements.txt
Flask==2.3.2
SQLAlchemy==2.0.21
web3==6.7.0
py-solc-x==1.3.0
qrcode==7.4
pillow==10.0.1
eth-account==0.9.0
eth-utils==2.3.0
